requirements.txt
altair @ file:///tmp/build/80754af9/altair_1599835197802/work
appdirs==1.4.4
attrs @ file:///C:/miniconda3/conda-bld/attrs_1762356943133/work
beautifulsoup4 @ file:///C:/miniconda3/conda-bld/beautifulsoup4-split_1764159464919/work
blinker @ file:///C:/miniconda3/conda-bld/blinker_1764332307774/work
Bottleneck @ file:///C:/miniconda3/conda-bld/bottleneck_1761938212172/work
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1764961392766/work
cachetools @ file:///C:/miniconda3/conda-bld/cachetools_1764867585295/work
certifi @ file:///C:/miniconda3/conda-bld/certifi_1763135348707/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832793693/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744988796/work
click @ file:///C:/miniconda3/conda-bld/click_1764332357576/work
colorama @ file:///C:/Users/dev-admin/buildout/croot/colorama_1759952627215/work
entrypoints @ file:///C:/Users/dev-admin/buildout/croot/entrypoints_1759980387324/work
frozendict @ file:///C:/miniconda3/conda-bld/frozendict_1761750689033/work
gdeltdoc==1.12.0
gitdb @ file:///C:/Users/dev-admin/buildout/croot/gitdb_1759961742190/work
GitPython @ file:///C:/Users/dev-admin/buildout/croot/gitpython_1759964586670/work
html5lib @ file:///Users/ktietz/demo/mc3/conda-bld/html5lib_1629144453894/work
idna @ file:///C:/Users/dev-admin/buildout/croot/idna_1762551230734/work
Jinja2 @ file:///C:/Users/dev-admin/buildout/croot/jinja2_1761938590194/work
jsonschema @ file:///C:/miniconda3/conda-bld/jsonschema_1762459329751/work
jsonschema-specifications @ file:///C:/miniconda3/conda-bld/jsonschema-specifications_1762788613712/work
lxml @ file:///C:/miniconda3/conda-bld/lxml_1764009760018/work
MarkupSafe @ file:///C:/Users/dev-admin/buildout/croot/markupsafe_1759955189290/work
mkl-service==2.5.2
mkl_fft @ file:///C:/miniconda3/conda-bld/mkl_fft_1761592920453/work
mkl_random @ file:///C:/miniconda3/conda-bld/mkl_random_1761593081220/work
multitasking @ file:///C:/Users/dev-admin/buildout/croot/multitasking_1760005137712/work
numexpr @ file:///C:/miniconda3/conda-bld/numexpr_1762165726164/work
numpy @ file:///C:/miniconda3/conda-bld/numpy_and_numpy_base_1763980696655/work/dist/numpy-2.3.5-cp314-cp314-win_amd64.whl#sha256=76af5c26e1b6f0492c90dbb17276bd664bbde95b360641a0da098aef53933ec8
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049098810/work
pandas @ file:///C:/miniconda3/conda-bld/pandas_1762332435977/work/dist/pandas-2.3.3-cp314-cp314-win_amd64.whl#sha256=ac6b8c30836d425b7de9199ad67d79b5d4d36663122fd5a70d81bfc894c0da42
peewee @ file:///C:/Users/dev-admin/buildout/croot/peewee_1761996257574/work
pillow @ file:///C:/Users/dev-admin/buildout/croot/pillow_1762547797675/work
platformdirs @ file:///C:/miniconda3/conda-bld/platformdirs_1762356618739/work
protobuf @ file:///C:/Users/dev-admin/buildout/croot/protobuf_1759959508725/work/bazel-bin/python/dist/protobuf-5.29.3-cp314-abi3-win_amd64.whl#sha256=19b451424e22e634c6c472582ee0de8154b8b411d6cd105050f14ee8acf353ab
pyarrow @ file:///C:/miniconda3/conda-bld/pyarrow_1764939513622/work/python
pycparser @ file:///C:/Users/dev-admin/buildout/croot/pycparser_1759955361075/work
pydeck @ file:///C:/Users/dev-admin/buildout/croot/pydeck_1761997772384/work
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753029698/work
python-dateutil @ file:///C:/Users/dev-admin/buildout/croot/python-dateutil_1759954839763/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dotenv_1761503229/work
pytz @ file:///C:/Users/dev-admin/buildout/croot/pytz_1759958332927/work
PyYAML @ file:///C:/miniconda3/conda-bld/pyyaml_1763465514484/work
referencing @ file:///C:/miniconda3/conda-bld/referencing_1762536906066/work
requests @ file:///C:/miniconda3/conda-bld/requests_1762359708170/work
rpds-py @ file:///C:/miniconda3/conda-bld/rpds-py_1762366636411/work
six @ file:///C:/Users/dev-admin/buildout/croot/six_1759954493303/work
smmap @ file:///tmp/build/80754af9/smmap_1611694433573/work
soupsieve @ file:///C:/Users/dev-admin/buildout/croot/soupsieve_1761938454850/work
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1764964602431/work
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1743640740204/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
toolz @ file:///C:/miniconda3/conda-bld/toolz_1764677750020/work
tornado @ file:///C:/Users/dev-admin/buildout/croot/tornado_1759954992390/work
typing_extensions @ file:///C:/Users/dev-admin/buildout/croot/typing_extensions_1759955056205/work
tzdata @ file:///croot/python-tzdata_1746123641790/work
urllib3 @ file:///C:/miniconda3/conda-bld/urllib3_1764965003544/work
vaderSentiment @ file:///home/conda/feedstock_root/build_artifacts/vadersentiment_1735583311926/work
watchdog @ file:///C:/Users/dev-admin/buildout/croot/watchdog_1759968437007/work
webencodings @ file:///C:/Users/dev-admin/buildout/croot/webencodings_1761938483125/work
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746275387/work
yfinance @ file:///home/conda/feedstock_root/build_artifacts/yfinance_1745889389907/work

# Google Sheets
gspread
google-auth
